This file is pointless
This file will needlessly grow in size and adding pictures is made harder. It is better to use the names of the files in the collection directory. If the need arises to map the same name to multiple different images, we can append numbers to the ends of the image names.